+++ test/style/module_directives_test.exs
@@ -468,4 +468,24 @@ defmodule Styler.Style.ModuleDirectivesTest do
       """
     )
   end
+
+  test "@derive movements" do
+    assert_style """
+                 defmodule F do
+                   defstruct [:a]
+                   @derive Inspect
+                   @derive {Foo, bar: :baz}
+                 end
+                 """,
+                 """
+                 defmodule F do
+                   @moduledoc false
+                   @derive Inspect
+                   @derive {Foo, bar: :baz}
+                   defstruct [:a]
+                 end
+                 """
+
+    assert_style "@derive Inspect"
+  end
 end